DEV-2128 Increase Chokidar parameter
Due to unstable connections, some files are marked as complete although
they are still being uploaded. There is a Chokidar parameter
`awaitWriteFinish.stabilityThreshold` which dictates the amount of time
in milliseconds for a file size to remain constant before emitting its event.

We increase that value to 10 seconds in order to counter that instability.